Apply code formatting markup to README.md
As it currently stands, the README.md file has indents and such to indicate a bunch of code.  While this is fine and dandy, when this renders in github, it does not display as 'code formatting' and therefore can be confusing if people read the README through Github instead of in a text editor.

This can be argued either way, but for aesthetics purposes, I thought I'd propose making the markup a little more code formatted to match other code formatting done in this document.