Say hello to a channel filter module for identified clients
The channel filter module maintains client specific channel lists for
identified clients. A typical use case is to have a subset of channels
visible for a mobile client.

Managing identified clients:

    /msg *chanfilter addclient <identifier>
    /msg *chanfilter delclient <identifier>
    /msg *chanfilter listclients

When an identified client connects ZNC first time, all channels on that
network are joined. The list of channels is automatically updated when
the identified client joins and parts channels. Next time the identified
client connects, it joins the channels it had visible from the last
session.