travis: only send success emails when state changes
I think everyone is on the channel anyway so it's not necressary to
receive emails every time the build succeeds. With this change emails
are sent:

* always on build failure.
* when the build is fixed again.

I would personally set `email: false`, but I think it was discussed
before somewhere related to ZNC and wasn't liked.

I am PRing this to 1.6.x, because pull requests also come there and it's
merged into master sooner or later.

As `travis lint` passes,

[CI SKIP]