Fix #811: Undernet channel key issue
Force CChan to be created so that a key passed in the JOIN command is saved.
Otherwise, if CChan didn't exist at this point and we weren't able to save
the key, it would be lost in case the server later presented the key as '*'
when ZNC requested the channel modes.