.travis.yml: enable fast_finish
When fast_finish is True, Travis will inform that the build has failed
when one build fails which is usually desirable instead of waiting for
everything to finish.

However builds that haven't finished yet will continue, this doesn't
stop them or do anything else to them.

`travis lint` passes so,

[CI SKIP]