Remove awaynick module from /modules
This module has been deprecated for a year now, and no longer serves an useful function in ZNC

Previously, the code would instruct ZNC to load simple away instead if awaynick was already loaded, and instruct users of its deprecation when tried to load.
But seeing as time has gone by, I think now is the time it would be fair to state
'this module cannot be found' over
'this module is deprecated - see this wiki page' or 'we're loading something else for you.'

It simply removes a bit of boilerplate code that in itself does not really do anything.

With the next release from znc still being far in sight, it might not be unreasonable to suggest this module can be safely removed at this point in time.