Fix #1133: Add MinClients option to simple_away
This adds a new option and command to the module, "MinClients", that
specifies the minimum number of clients that must be connected for the
module to considered the user as available.  If fewer clients are
connected than the minimum, the user will be marked away, even if there
are still clients attached.  The default value is "1", which replicates
the existing behavior of the module.

To modify the MinClients value, `/msg *simple_away MinClients [<value>]`
is available.  Running the command without parameters will display the
current value instead of modifying it.