Change default flood rates to match 99% of IRC networks settings, prevent excess flood problems
I noticed that the default network flood rates are wrong (for most networks). It has been very standard since the beginning of IRC to use a value of 2 seconds per line, with a 10 line head start. Almost all networks work this way, and if you don't follow, having a lot of channels or a lot of clients connected to ZNC results in excess flood disconnections and much frustration for the user. 

I think you should make the default value for these 2 and 9 instead of 1 and 4 so that this nasty bug doesn't hit people who will not know how to debug it. You find these same types of settings built into every IRC client. 2 and 10 is the standard. 2 and 9 provides a slight buffer so there is no mistake.

Thanks
-Rubin