Make some changes to sample module
***This module was written as an example, and should be treated as such***.

It details the various types of things that a module can do; however it is not really a "true" end-user-usable module in that it's not really designed to be "professionally written" like the rest of the modules.

To that end, #1511 tried to make it more professional, but removed a bit of code that was supposed to be kept.  This was done intentionally so that you could show examples of how modules could be used.

This commit fixes the issues in 1511 but doesn't remove the "RealName" suffix appending example.  It instead puts " - Running on ZNC" as the example string.