Use a separate lock file instead of locking znc.conf
Instead of locking the config file, open a `<config>.lock` file
and put the fcntl lock on it.

Before this change, the config file needed to be opened read-write,
even when we only want to read it, because of fcntl locks.

The locks are meant to prevent two ZNCs from running on the same config,
when they could overwrite each others' config changes.
However, since any configs writes and rehashes always reopen the config
by path, it's the path to the config that we want to lock,
not the file itself
(eg. if someone moves znc.conf to a different directory and creates
a new one in our directory, we still don't want another znc instance
to run using the new config, but we don't mind another znc instance
running using the old config in its new location).

Therefore, a separate lock file is not worse than locking the config file
itself, and it will allow to open the config read-only.