Add an option to run in read-only mode
Sometimes it is desirable for ZNC not to overwrite its config,
eg. because the config comes from external configuration management tools.

Add a commandline option to set ZNC in read-only mode, in which
it will open the config file read-only, and will refuse to overwrite it.
Note that this does not affect .registry files, which can still be written.

Also add a big fat warning in the console when running in read-only mode.