modules: Remove Q
Q can be replaced with the perform module, and will still be available
as a third-party module for those who use the module and QuakeNet.

In general, the Q module only targets a single network not everyone may
use, the additional support and maintenance burden is not worth it.

References #786
Closes #1636 as wontfix
Closes #554 as wontfix