Add account to joins for the log module.
This commit adds the account name for identified users to "Joins" lines in logs
generated by the log module.  It can get the account name from either the
account tag (if the account-tag capability is requested) or the extended-join
info (if the extended-join capability is requested).  The current version of ZNC
requests both, but this feature will still work if the IRC server only supports
one or the other.