docker: Fix startup-sequence locations
In https://github.com/znc/znc-docker/pull/28 the location of the init
scripts was moved which broke the logic to copy all the init files once
the docker submodule was updated in
5d7aa0f7d44cb0d686881ced5b409c3d0569e7ca. Consequently only
`30-build-modules.sh` was available in the build docker image. We fix
this issue by also including all startup-sequence subfolders.

Signed-off-by: Christian Heusel <christian@heusel.eu>