Allow `--makepass` to work non-interactively
When automating deployments of ZNC, it can be helpful to capture a
generated '<Password>' section. Allow users to do this by defining an
environment variable `ZNC_PASS` containing the chosen password.

The password block can be retrieved with the following snippet when the
environment variable is set;

```
znc --makepass | sed -n '/Pass/,/Pass/p'
```