Make hostname detection a requirement for certificate creation
If the HOSTNAME environment variable is not defined while generating
the self-signed certificate, the certificate Common Name will be silently
defined as "host.unknown":

    $ (unset HOSTNAME && ./build/src/znc -d $PWD/badcert -p)
    $ openssl x509 -in ./badcert/znc.pem -inform PEM -text | grep unknown
        Issuer: OU=dacav, CN=host.unknown, emailAddress=dacav@host.unknown
        Subject: OU=dacav, CN=host.unknown, emailAddress=dacav@host.unknown

Unless "host.unknown" is matching the actual host name (unlikely), the obtained
certificate will not be acceptable, even if marked as trusted by the client.

This patch makes the certificate generation fail if the host name can not be
determined by reading the HOSTNAME environment variable.